make childViews synchronized for client/server mode
In the old version, the list of child view defs was synchronized. That was
fed into the repeat attribute with a repeatWrapper that would create the
ProductView, ContentView, SlideshowView etc. But, we don't have a convenient
way to synchronize the repeatTags in an element. It's possible that we could
add sync=true for tag objects and when that's set on a repeat we sync the
repeat tags.

Instead, with this checkin the ParentDef class creates the list of child
views stored as a property of the PageView. That property is synchronized
and is used in the repeat attribute. The repeat wrapper just returns the
sets the parent node, id, and returns the repeat element -
so repeatTags = repeat essentially. This works for the most part but need
to make sure repeatTags are not destroyed when they are removed (since they
were not created in the first place as part of that createElement operation).